Add option to automatically simplify branch geometry
When exported to an FBX or OBJ, curves are automatically converted to
meshes. The way that Blender does this, however, results in tons of
unnecessary geometries. This PR provides an option to do this process in
a more optimized manner automatically.

First, the branches curve is converted to a mesh. Second, a limited
dissolve is run on the mesh to remove unneeded faces / verts / etc.
These two steps are not terribly difficult to do on one's own, but
including this option is a great QOL improvement in my opinion.

---
Here is a comparison of a FBX exports of a Quaking Aspen model:

Curves filesize: 38.1 MB

[Curves converted to mesh](https://i.imgur.com/D40ElEF.jpg)
867,624 verts | 839,760 faces | 1,679,520 tris
Filesize: 38.1 MB

[Simplified branch geometry](https://i.imgur.com/WEAUONb.jpg) -- [mesh
view](https://i.imgur.com/P09dN8n.jpg)
546,102 verts | 254,653 faces | 724,716 tris
Filesize: 18.8 MB

Savings:
321,522 verts | 585,107 faces | 954,804 tris