Fix for crash when running LOD creation automatically; memory handling
LOD creation is now run on the main (GUI) thread.

Applying modifiers (in this case decimate) does not play well with threading (`bpy` is not thread-safe, so much of what we do is a little sketchy to begin with), so I added a modal that allows the addon's thread to queue tasks for running on the main thread.

I also added some code to get every last bit of temporary data out of memory ASAP.